# Small optimizations