Three smaller optimizations:
- Track current program, only update if it changes
- Check for empty objects passed to `setParameters`
- Keep track of texture uniforms in `Program` so you don't have to loop through all uniforms to bind them (deck's tends to send a lot of uniforms, not many textures for rendering).

Total performance improvement over 7.3 on the stress test currently at:

|        |7.3  |Current|Improvement|
|--------|-----|-------|-----------|        
|CPU Time|43ms |17ms   |26ms (60%) |
|GPU Time|38ms |24ms   |14ms (37%) |